Handle matrix layout decorations
This is a bit ugly, since matrix layout decorations are applied to
structure members, not to the actual matrix and vector pointer types
that they affect. We essentially have to track the matrix layout in the
object that is produced by an OpAccessChain, and implement special-case
handling when loading and storing with these objects.

Also combined the OpAccessChain and OpPtrAccessChain implementations to
avoid excessive duplicated code.